packages/ai/src/ai/common/models/ocr/requirements_trocr.txt
# Requirements for TrOCR loader (detection + recognition)
# Direct imports only
# opencv is handled by ai.common.opencv module

transformers
craft-text-detector
Pillow
